LayoutTests/platform/glib/accessibility/multiple-label-input-expected.txt

This test ensures that inputs with multiple labels have the correct accessibility text exposed.

Initial AX text:
	AXTitle: Enter your color
	AXDescription:
document.getElementById('label2').remove();
	AXTitle: Enter color
	AXDescription:
document.getElementById('label3').innerText = 'choice';
	AXTitle: Enter choice
	AXDescription:
document.getElementById('label3').innerHTML = 'decision';
	AXTitle: Enter decision
	AXDescription:

PASS successfullyParsed is true

TEST COMPLETE
 Enter decision
